chore: rename fetch_item_thumbnails
